suzuki-technology-trends/backend/pyproject.toml
[project]
name = "drawer-ai-backend"
version = "2.0.0"
description = "DrawerAI Technology Review System - LangGraph Backend"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    # LangGraph
    "langgraph>=1.0.0",
    # LangChain
    "langchain>=0.3.0",
    "langchain-google-genai>=4.0.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    # HTTP Client
    "httpx>=0.25.0",
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    # Async
    "nest-asyncio>=1.5.8",
    # Text Processing
    "markdown>=3.4.0",
    # Environment
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    # Other
    "python-dateutil>=2.8.2",
    "typing_extensions>=4.0.0",
    # Authentication
    "PyJWT>=2.8.0",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1", "pytest>=7.4.0", "pytest-asyncio>=0.21.0"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
]
lint.ignore = [
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.11",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
